Break up websocket component (#17003)
* Break up websocket component

* Lint
