+++ homeassistant/components/websocket_api/__init__.py
@@ -7,7 +7,7 @@ https://developers.home-assistant.io/docs/external_api_websocket.html
 import asyncio
 from concurrent import futures
 from contextlib import suppress
-from functools import partial, wraps
+from functools import partial
 import json
 import logging
 
@@ -15,19 +15,17 @@ from aiohttp import web
 import voluptuous as vol
 from voluptuous.humanize import humanize_error
 
-from homeassistant.const import (
-    MATCH_ALL, EVENT_TIME_CHANGED, EVENT_HOMEASSISTANT_STOP,
-    __version__)
-from homeassistant.core import Context, callback, HomeAssistant
+from homeassistant.const import EVENT_HOMEASSISTANT_STOP, __version__
+from homeassistant.core import Context, callback
 from homeassistant.loader import bind_hass
 from homeassistant.helpers.json import JSONEncoder
-from homeassistant.helpers import config_validation as cv
-from homeassistant.helpers.service import async_get_all_descriptions
 from homeassistant.components.http import HomeAssistantView
 from homeassistant.components.http.auth import validate_password
 from homeassistant.components.http.const import KEY_AUTHENTICATED
 from homeassistant.components.http.ban import process_wrong_login, \
     process_success_login
+
+from . import commands, const, decorators, messages
 
 DOMAIN = 'websocket_api'
 
@@ -36,85 +34,30 @@ DEPENDENCIES = ('http',)
 
 MAX_PENDING_MSG = 512
 
-ERR_ID_REUSE = 1
-ERR_INVALID_FORMAT = 2
-ERR_NOT_FOUND = 3
-ERR_UNKNOWN_COMMAND = 4
-ERR_UNKNOWN_ERROR = 5
-
-TYPE_AUTH = 'auth'
-TYPE_AUTH_INVALID = 'auth_invalid'
-TYPE_AUTH_OK = 'auth_ok'
-TYPE_AUTH_REQUIRED = 'auth_required'
-TYPE_CALL_SERVICE = 'call_service'
-TYPE_EVENT = 'event'
-TYPE_GET_CONFIG = 'get_config'
-TYPE_GET_SERVICES = 'get_services'
-TYPE_GET_STATES = 'get_states'
-TYPE_PING = 'ping'
-TYPE_PONG = 'pong'
-TYPE_RESULT = 'result'
-TYPE_SUBSCRIBE_EVENTS = 'subscribe_events'
-TYPE_UNSUBSCRIBE_EVENTS = 'unsubscribe_events'
 
 _LOGGER = logging.getLogger(__name__)
 
 JSON_DUMP = partial(json.dumps, cls=JSONEncoder)
 
+TYPE_AUTH = 'auth'
+TYPE_AUTH_INVALID = 'auth_invalid'
+TYPE_AUTH_OK = 'auth_ok'
+TYPE_AUTH_REQUIRED = 'auth_required'
+
+
+# Backwards compat
+# pylint: disable=invalid-name
+BASE_COMMAND_MESSAGE_SCHEMA = messages.BASE_COMMAND_MESSAGE_SCHEMA
+error_message = messages.error_message
+result_message = messages.result_message
+async_response = decorators.async_response
+ws_require_user = decorators.ws_require_user
+# pylint: enable=invalid-name
+
 AUTH_MESSAGE_SCHEMA = vol.Schema({
     vol.Required('type'): TYPE_AUTH,
     vol.Exclusive('api_password', 'auth'): str,
     vol.Exclusive('access_token', 'auth'): str,
-})
-
-# Minimal requirements of a message
-MINIMAL_MESSAGE_SCHEMA = vol.Schema({
-    vol.Required('id'): cv.positive_int,
-    vol.Required('type'): cv.string,
-}, extra=vol.ALLOW_EXTRA)
-# Base schema to extend by message handlers
-BASE_COMMAND_MESSAGE_SCHEMA = vol.Schema({
-    vol.Required('id'): cv.positive_int,
-})
-
-
-SCHEMA_SUBSCRIBE_EVENTS = BASE_COMMAND_MESSAGE_SCHEMA.extend({
-    vol.Required('type'): TYPE_SUBSCRIBE_EVENTS,
-    vol.Optional('event_type', default=MATCH_ALL): str,
-})
-
-
-SCHEMA_UNSUBSCRIBE_EVENTS = BASE_COMMAND_MESSAGE_SCHEMA.extend({
-    vol.Required('type'): TYPE_UNSUBSCRIBE_EVENTS,
-    vol.Required('subscription'): cv.positive_int,
-})
-
-
-SCHEMA_CALL_SERVICE = BASE_COMMAND_MESSAGE_SCHEMA.extend({
-    vol.Required('type'): TYPE_CALL_SERVICE,
-    vol.Required('domain'): str,
-    vol.Required('service'): str,
-    vol.Optional('service_data'): dict
-})
-
-
-SCHEMA_GET_STATES = BASE_COMMAND_MESSAGE_SCHEMA.extend({
-    vol.Required('type'): TYPE_GET_STATES,
-})
-
-
-SCHEMA_GET_SERVICES = BASE_COMMAND_MESSAGE_SCHEMA.extend({
-    vol.Required('type'): TYPE_GET_SERVICES,
-})
-
-
-SCHEMA_GET_CONFIG = BASE_COMMAND_MESSAGE_SCHEMA.extend({
-    vol.Required('type'): TYPE_GET_CONFIG,
-})
-
-
-SCHEMA_PING = BASE_COMMAND_MESSAGE_SCHEMA.extend({
-    vol.Required('type'): TYPE_PING,
 })
 
 
@@ -148,46 +91,6 @@ def auth_invalid_message(message):
     }
 
 
-def event_message(iden, event):
-    """Return an event message."""
-    return {
-        'id': iden,
-        'type': TYPE_EVENT,
-        'event': event.as_dict(),
-    }
-
-
-def error_message(iden, code, message):
-    """Return an error result message."""
-    return {
-        'id': iden,
-        'type': TYPE_RESULT,
-        'success': False,
-        'error': {
-            'code': code,
-            'message': message,
-        },
-    }
-
-
-def pong_message(iden):
-    """Return a pong message."""
-    return {
-        'id': iden,
-        'type': TYPE_PONG,
-    }
-
-
-def result_message(iden, result=None):
-    """Return a success result message."""
-    return {
-        'id': iden,
-        'type': TYPE_RESULT,
-        'success': True,
-        'result': result,
-    }
-
-
 @bind_hass
 @callback
 def async_register_command(hass, command, handler, schema):
@@ -198,43 +101,10 @@ def async_register_command(hass, command, handler, schema):
     handlers[command] = (handler, schema)
 
 
-def require_owner(func):
-    """Websocket decorator to require user to be an owner."""
-    @wraps(func)
-    def with_owner(hass, connection, msg):
-        """Check owner and call function."""
-        user = connection.request.get('hass_user')
-
-        if user is None or not user.is_owner:
-            connection.to_write.put_nowait(error_message(
-                msg['id'], 'unauthorized', 'This command is for owners only.'))
-            return
-
-        func(hass, connection, msg)
-
-    return with_owner
-
-
 async def async_setup(hass, config):
     """Initialize the websocket API."""
     hass.http.register_view(WebsocketAPIView)
-
-    async_register_command(hass, TYPE_SUBSCRIBE_EVENTS,
-                           handle_subscribe_events, SCHEMA_SUBSCRIBE_EVENTS)
-    async_register_command(hass, TYPE_UNSUBSCRIBE_EVENTS,
-                           handle_unsubscribe_events,
-                           SCHEMA_UNSUBSCRIBE_EVENTS)
-    async_register_command(hass, TYPE_CALL_SERVICE,
-                           handle_call_service, SCHEMA_CALL_SERVICE)
-    async_register_command(hass, TYPE_GET_STATES,
-                           handle_get_states, SCHEMA_GET_STATES)
-    async_register_command(hass, TYPE_GET_SERVICES,
-                           handle_get_services, SCHEMA_GET_SERVICES)
-    async_register_command(hass, TYPE_GET_CONFIG,
-                           handle_get_config, SCHEMA_GET_CONFIG)
-    async_register_command(hass, TYPE_PING,
-                           handle_ping, SCHEMA_PING)
-
+    commands.async_register_commands(hass)
     return True
 
 
@@ -389,19 +259,19 @@ class ActiveConnection:
 
             while msg:
                 self.debug("Received", msg)
-                msg = MINIMAL_MESSAGE_SCHEMA(msg)
+                msg = messages.MINIMAL_MESSAGE_SCHEMA(msg)
                 cur_id = msg['id']
 
                 if cur_id <= last_id:
-                    self.to_write.put_nowait(error_message(
-                        cur_id, ERR_ID_REUSE,
+                    self.to_write.put_nowait(messages.error_message(
+                        cur_id, const.ERR_ID_REUSE,
                         'Identifier values have to increase.'))
 
                 elif msg['type'] not in handlers:
                     self.log_error(
                         'Received invalid command: {}'.format(msg['type']))
-                    self.to_write.put_nowait(error_message(
-                        cur_id, ERR_UNKNOWN_COMMAND,
+                    self.to_write.put_nowait(messages.error_message(
+                        cur_id, const.ERR_UNKNOWN_COMMAND,
                         'Unknown command.'))
 
                 else:
@@ -410,8 +280,8 @@ class ActiveConnection:
                         handler(self.hass, self, schema(msg))
                     except Exception:  # pylint: disable=broad-except
                         _LOGGER.exception('Error handling message: %s', msg)
-                        self.to_write.put_nowait(error_message(
-                            cur_id, ERR_UNKNOWN_ERROR,
+                        self.to_write.put_nowait(messages.error_message(
+                            cur_id, const.ERR_UNKNOWN_ERROR,
                             'Unknown error.'))
 
                 last_id = cur_id
@@ -435,8 +305,8 @@ class ActiveConnection:
                 else:
                     iden = None
 
-                final_message = error_message(
-                    iden, ERR_INVALID_FORMAT, error_msg)
+                final_message = messages.error_message(
+                    iden, const.ERR_INVALID_FORMAT, error_msg)
 
         except TypeError as err:
             if wsock.closed:
@@ -485,170 +355,3 @@ class ActiveConnection:
             self.debug("Closed connection")
 
         return wsock
-
-
-def async_response(func):
-    """Decorate an async function to handle WebSocket API messages."""
-    async def handle_msg_response(hass, connection, msg):
-        """Create a response and handle exception."""
-        try:
-            await func(hass, connection, msg)
-        except Exception:  # pylint: disable=broad-except
-            _LOGGER.exception("Unexpected exception")
-            connection.send_message_outside(error_message(
-                msg['id'], 'unknown', 'Unexpected error occurred'))
-
-    @callback
-    @wraps(func)
-    def schedule_handler(hass, connection, msg):
-        """Schedule the handler."""
-        hass.async_create_task(handle_msg_response(hass, connection, msg))
-
-    return schedule_handler
-
-
-@callback
-def handle_subscribe_events(hass, connection, msg):
-    """Handle subscribe events command.
-
-    Async friendly.
-    """
-    async def forward_events(event):
-        """Forward events to websocket."""
-        if event.event_type == EVENT_TIME_CHANGED:
-            return
-
-        connection.send_message_outside(event_message(msg['id'], event))
-
-    connection.event_listeners[msg['id']] = hass.bus.async_listen(
-        msg['event_type'], forward_events)
-
-    connection.to_write.put_nowait(result_message(msg['id']))
-
-
-@callback
-def handle_unsubscribe_events(hass, connection, msg):
-    """Handle unsubscribe events command.
-
-    Async friendly.
-    """
-    subscription = msg['subscription']
-
-    if subscription in connection.event_listeners:
-        connection.event_listeners.pop(subscription)()
-        connection.to_write.put_nowait(result_message(msg['id']))
-    else:
-        connection.to_write.put_nowait(error_message(
-            msg['id'], ERR_NOT_FOUND, 'Subscription not found.'))
-
-
-@async_response
-async def handle_call_service(hass, connection, msg):
-    """Handle call service command.
-
-    Async friendly.
-    """
-    blocking = True
-    if (msg['domain'] == 'homeassistant' and
-            msg['service'] in ['restart', 'stop']):
-        blocking = False
-    await hass.services.async_call(
-        msg['domain'], msg['service'], msg.get('service_data'), blocking,
-        connection.context(msg))
-    connection.send_message_outside(result_message(msg['id']))
-
-
-@callback
-def handle_get_states(hass, connection, msg):
-    """Handle get states command.
-
-    Async friendly.
-    """
-    connection.to_write.put_nowait(result_message(
-        msg['id'], hass.states.async_all()))
-
-
-@async_response
-async def handle_get_services(hass, connection, msg):
-    """Handle get services command.
-
-    Async friendly.
-    """
-    descriptions = await async_get_all_descriptions(hass)
-    connection.send_message_outside(
-        result_message(msg['id'], descriptions))
-
-
-@callback
-def handle_get_config(hass, connection, msg):
-    """Handle get config command.
-
-    Async friendly.
-    """
-    connection.to_write.put_nowait(result_message(
-        msg['id'], hass.config.as_dict()))
-
-
-@callback
-def handle_ping(hass, connection, msg):
-    """Handle ping command.
-
-    Async friendly.
-    """
-    connection.to_write.put_nowait(pong_message(msg['id']))
-
-
-def ws_require_user(
-        only_owner=False, only_system_user=False, allow_system_user=True,
-        only_active_user=True, only_inactive_user=False):
-    """Decorate function validating login user exist in current WS connection.
-
-    Will write out error message if not authenticated.
-    """
-    def validator(func):
-        """Decorate func."""
-        @wraps(func)
-        def check_current_user(hass: HomeAssistant,
-                               connection: ActiveConnection,
-                               msg):
-            """Check current user."""
-            def output_error(message_id, message):
-                """Output error message."""
-                connection.send_message_outside(error_message(
-                    msg['id'], message_id, message))
-
-            if connection.user is None:
-                output_error('no_user', 'Not authenticated as a user')
-                return
-
-            if only_owner and not connection.user.is_owner:
-                output_error('only_owner', 'Only allowed as owner')
-                return
-
-            if (only_system_user and
-                    not connection.user.system_generated):
-                output_error('only_system_user',
-                             'Only allowed as system user')
-                return
-
-            if (not allow_system_user
-                    and connection.user.system_generated):
-                output_error('not_system_user', 'Not allowed as system user')
-                return
-
-            if (only_active_user and
-                    not connection.user.is_active):
-                output_error('only_active_user',
-                             'Only allowed as active user')
-                return
-
-            if only_inactive_user and connection.user.is_active:
-                output_error('only_inactive_user',
-                             'Not allowed as active user')
-                return
-
-            return func(hass, connection, msg)
-
-        return check_current_user
-
-    return validator
