+++ homeassistant/components/websocket_api/const.py
@@ -1,0 +1,8 @@
+"""Websocket constants."""
+ERR_ID_REUSE = 1
+ERR_INVALID_FORMAT = 2
+ERR_NOT_FOUND = 3
+ERR_UNKNOWN_COMMAND = 4
+ERR_UNKNOWN_ERROR = 5
+
+TYPE_RESULT = 'result'
