+++ tests/components/websocket_api/__init__.py
@@ -1,0 +1,2 @@
+"""Tests for the websocket API."""
+API_PASSWORD = 'test1234'
